requirements.txt
click==8.0.4
Flask==2.0.3
Flask-WTF==1.0.1
gunicorn==20.1.0
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
PyJWT==2.3.0
python-dotenv==0.19.2
Werkzeug==2.0.3
WTForms==3.0.1
Flask-Login
PyJWT
Flask-SQLAlchemy
psycopg2
Flask-Migrate